fix func signature, includes strings, fix nanosleep

+++ include/message.h
@@ -31,7 +31,7 @@ void share_peers_list(peer server, int *clock, pthread_mutex_t *clock_lock, peer
 char *build_message(sockaddr_in sender_ip, int clock, MSG_TYPE msg_type, void *args);                   // create a message with the sender ip, its clock and the message type
 void send_message(const char *msg, peer *neighbour, MSG_TYPE msg_type);                                 // send a built message to an known peer socket
 MSG_TYPE read_message(const char *buf, int *clock, peer *sender);                                       // read message, mark its sender and return the message type
-char *check_msg_full(const char *buf, SOCKET sock, int *rec_peers_size, int *valread);                  // check if message received was read fully
+char *check_msg_full(const char *buf, SOCKET sock, int *rec_peers_size, ssize_t *valread);                  // check if message received was read fully
 void append_list_peers(const char *buf, peer **peers, size_t *peers_size, size_t rec_peers_size, char *file);       // append received list to known peer list
 void bye_peers(peer server, int *clock, peer *peers, size_t peers_size);                                // send a bye message to every peer in list
